Changes to SoundManager, Added Shoot Sound
Slightly altered functionality of SoundManager: now has two separate audio sources, one for clips which can overlap and should play to the end no matter what, and one for clips which should not overlap and are mutually exclusive. These are utilized via two separate static methods, known as PlaySoundContinuous and PlaySoundInterrupt, respectively.

Shock/stun/hit sound for both Normal and Unstable Bats now uses the interruptSource, meaning only one instance of the hit sound may be played at once.

Also added a sound for shooting, which is played through the InputManager on click (when the game is not paused).

Can be tested with mouse. Test by playing the game, and noting how/when/which sounds overlap or cut each other off. Also try this with the Title screen, because I haven't done anything with it [directly] just yet.

Due in Sprint 4.