added menu click sound effect
it's a bit scuffed to listen to with headphones, and it's very quiet, but a menu click sound has been added.

Updated all relevant Scene entities and scripts to use the new menu sound.

GameOverBehavior now sets the timeScale to 0 upon loading the Game Over screen.

InputManager now only plays the Shoot sound when the timeScale is >0.

Shoot sound removed from the Title Menu.

Can test with mouse.
Click menu buttons to test.

Due in sprint 5.