very minor change to the flame presentation on the mod bat sprite
Made it so that the plumes at the start and puff near the end don't squish into the border.

Saved this as a separate version, as it doesn't seem to make that big of a difference. I don't plan on implementing this new version into the build any time soon, though I want to keep it around just in case.