it lives
for the love of god, please let me sleep

in all seriousness, while getting used to illustrating was an exercise in frustration, I feel it was worth it to try illustrating for the first time.

The project isn't finished, mind you, but I'm most of the way there (this time for sure (<-- isn't sure at all  ;u;)).