added alternate anims for divebomb bat
new versions have a slight border on the wing end of the third frame. This is tan attempt to prevent the sprites from looking "cut off" due to the positioning of their arm and wing.

No alt is made for the Stun anim, as this one lacks the relevant pose (wing folded halfway).