minor tweaks
renamed a private property of AnimationHandler to fall in line with coding standards.

changed animation components on bat prefabs to have the correct default animations. NOTE: I don't think this actually has any impact on anything since we're using the animation controller/state machine, but I wanted to do this for the sake of neatness and consistency.