Create powerCore_recolor.aseprite
began work on recolor for the power core sprite.
that is, the recolor is finished, pending teammate approval. Since the task was meant to be Lucas' (and I stole it out from under him), I'm probably going to let him handle implementing it, or maybe remaking it if the current look is not to his liking.