Refactored Music ScriptableObject, Implemented
implemented a new MusicTrack ScriptableObject, which stores AudioClips and delay offsets for music files.
This was done to account for a 100ms delay to the start and 250ms tail at the end of the persistent BGM's music.

Implemented the seamless looping and playback of the BGM -- the intro and the loop -- according to the declared StartOffset and EndOffset values.

Made it so that PlayNonloopMusic() will mute all musicLoopSources while playing.

Made minor changes to SoundManager to bring it more in line with Coding Standards

(sorry this one is so massive: it's tough to disentangle these changes)