Reorganized Archive, Recolored some Sprites
Reorganized the Art Archive base dir so that it's not quite as flooded.

Relocated the contents of the "misc" subdir to a local "samples" folder. Some deprecated/outdated items among these were simply deleted.

added recolor 1 for Unstable Bats, recolors 1 and 2 for Modbats, and aseprite file for recolor 3 of Modbats.

deleted an archaic (and frankly disgusting) WIP recolor of Modbats.

Fixed an issue in the Modbat's default palette's Attack animation: the fingers on frames 6 and 10 were not copied properly and were thus misaligned with the wingtips.